Stabilize AirSense AI: migrate Spring Security to 6.x lambda DSL, fix Redis auto-config, scale risk score to 0-100 integer, fix Leaflet map window.L markers, and improve ADK imports robustness

+++ backend-java/src/main/java/com/airsense/ai/model/RiskPrediction.java
@@ -26,7 +26,7 @@ public class RiskPrediction {
     private User user;
     
     @Column(name = "risk_score")
-    private double riskScore;
+    private int riskScore;
     
     @Column(name = "risk_level")
     private String riskLevel; // LOW, MEDIUM, HIGH
